+++ Tests/tuliPS.tests.ps1
@@ -94,35 +94,35 @@ Describe 'Commands' {
         Remove-Module $ModuleName
     }
 
-    Context 'Reset-OutDefault' {
+    # Context 'Reset-OutDefault' {
 
-        Reset-OutDefault
-        $cmd = Get-Command Out-Default -CommandType Cmdlet
+    #     Reset-OutDefault
+    #     $cmd = Get-Command Out-Default -CommandType Cmdlet
 
-        It 'Removes tuliPS Out-Default' {
-            Test-Path function:\Out-Default | Should Be $false
-        }
+    #     It 'Removes tuliPS Out-Default' {
+    #         Test-Path function:\Out-Default | Should Be $false
+    #     }
 
-        It 'Resets Out-Default' {
-            $cmd.Source | Should Be Microsoft.Powershell.Core
-        }
-    }
+    #     It 'Resets Out-Default' {
+    #         $cmd.Source | Should Be Microsoft.Powershell.Core
+    #     }
+    # }
 
-    Context 'Invoke-Wrapper' {
-        Invoke-Wrapper
+    # Context 'Invoke-Wrapper' {
+    #     Invoke-Wrapper
 
-        It 'Sets tuliPS Out-Default' {
-            $cmd = Get-Item function:\Out-Default 
-            Reset-OutDefault
-            $original = Get-Command Out-Default 
-            $original.Name | Should Be $cmd.Name
-            $original.CommandType | Should Not Be $cmd.CommandType 
-            Invoke-Wrapper 
-            $wrapped = Get-Command Out-Default
-            $wrapped.Name | Should Be $cmd.Name
-            $wrapped.CommandType | Should Be $cmd.CommandType
-        }
-    }
+    #     It 'Sets tuliPS Out-Default' {
+    #         $cmd = Get-Item function:\Out-Default 
+    #         Reset-OutDefault
+    #         $original = Get-Command Out-Default 
+    #         $original.Name | Should Be $cmd.Name
+    #         $original.CommandType | Should Not Be $cmd.CommandType 
+    #         Invoke-Wrapper 
+    #         $wrapped = Get-Command Out-Default
+    #         $wrapped.Name | Should Be $cmd.Name
+    #         $wrapped.CommandType | Should Be $cmd.CommandType
+    #     }
+    # }
 
     Context 'Get-Tulips' {
         InModuleScope tuliPS {
@@ -318,6 +318,44 @@ Describe 'Commands' {
         }
     }
 
+    Context 'Add-TulipsFormatter' {
+        InModuleScope tuliPS {
+            $sb = { Write-Host $_ -ForegroundColor Red } 
+            $newSb = { Write-Host $_ -ForegroundColor Blue } 
+
+            It 'Properly adds a formatter' { 
+                Add-TulipsFormatter -Type System.String -ScriptBlock $sb
+
+                $TypeTracker.ContainsKey('System.String') | Should Be $true
+                $TypeTracker.ContainsValue($sb) | Should Be $true
+            }
+
+            It 'Cannot set a formatter' {
+                { Add-TulipsFormatter -Type System.String -ScriptBlock $newSb$ } | Should Throw
+            }
+        } 
+    } 
+
+    Context 'Set-TulipsFormatter' {
+        InModuleScope tuliPS {
+            $sb = { Write-Host $_ -ForegroundColor Red } 
+            $newSb = { Write-Host $_ -ForegroundColor Blue } 
+
+            It 'Adds a formatter' { 
+                Set-TulipsFormatter -Type System.String -ScriptBlock $sb
+
+                $TypeTracker.ContainsKey('System.String') | Should Be $true
+                $TypeTracker.ContainsValue($sb) | Should Be $true
+            }
+
+            It 'Sets a formatter' {
+                Set-TulipsFormatter -Type System.String -ScriptBlock $newSb
+
+                $TypeTracker.ContainsKey('System.String') | Should Be $true
+                $TypeTracker.ContainsValue($newSb) | Should Be $true
+            }
+        }
+    }
 }
 
 # If the module was loaded but is not now, reload it