+++ changelog.md
@@ -2,7 +2,7 @@
 
 <endowdly@gmail.com>
 
-Versioning is controlled by Nebula.
+Versioning is controlled by Nova.
 To get the version, view the [version](./.version) file.
   
 ## [1.0.0] - 2019-03-04
@@ -15,7 +15,26 @@ To get the version, view the [version](./.version) file.
 - A full PowerShell module
 - Screenshots of tuliPS in action
 
+## [1.1.0] - 2019-03-12
+
+### Added
+
+- Add-TulipsFormatter.ps1: Add tuliPS formatting extensions!
+- Set-TulipsFormatter.ps1: Set tuliPS formatting extensions!
+- Get-ProcessBlock.ps1: An internal function that helps tuliPS format things.
+- Invoke-Wrapper.ps1: OutDefaultWrapper.ps1 reborn.
+
+### Removed
+
+- OutDefaultWrapper.ps1
+
+### Changed
+
+- Invoke-Wrapper.ps1: Now an internal function.
+- Reset-OutDefault.ps1: Now an internal function.
+- build/config.psd1: Reflected changes.
+
 ---
-Last Updated: 2019-03-04T07:51:21.4718579-05:00
+Last Updated: 2019-03-12T09:39:34.8506421-04:00
 
 <!-- ref -->